services/ingestor_finnhub/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.4.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0